# task/openapi.yml
openapi: "3.0.1"
info:
  title: Tender Management API
  version: "1.0"
  description: |
    API для управления тендерами и предложениями. 

    Основные функции API включают управление тендерами (создание, изменение, получение списка) и управление предложениями (создание, изменение, получение списка).
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер API
paths:
...
  /bids/new:
    post:
      summary: Создание нового предложения
      description: Создание предложения для существующего тендера.
      security:
        - bearerAuth: []
      operationId: createBid
      requestBody:
        description: Данные нового предложения.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/bidName"
                description:
                  $ref: "#/components/schemas/bidDescription"
                status:
                  $ref: "#/components/schemas/bidStatus"
                tenderId:
                  $ref: "#/components/schemas/tenderId"
                organizationId:
                  $ref: "#/components/schemas/organizationId"
                creatorUsername:
                  $ref: "#/components/schemas/username"
              required:
                - name
                - description
                - status
                - tenderId
                - organizationId
                - creatorUsername
      responses:
        "200":
          description: Предложение успешно создано. Сервер присваивает уникальный идентификатор и время создания.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
...
  /bids/{bidId}/status:
    get:
      summary: Получение текущего статуса предложения
      description: Получить статус предложения по его уникальному идентификатору.
      security:
        - bearerAuth: []
      operationId: getBidStatus
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Текущий статус предложения.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bidStatus"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
    put:
      summary: Изменение статуса предложения
      description: Изменить статус предложения по его уникальному идентификатору.
      operationId: updateBidStatus
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/bidStatus"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Статус предложения успешно изменен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Предложение не найдено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
...
  /bids/{bidId}/submit_decision:
    put:
      summary: Отправка решения по предложению
      description: Отправить решение (одобрить или отклонить) по предложению.
      operationId: submitBidDecision
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: decision
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/bidDecision"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Решение по предложению успешно отправлено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "400":
          description: Решение не может быть отправлено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Предложение не найдено.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
...
  /bids/{bidId}/rollback/{version}:
    put:
      summary: Откат версии предложения
      description: Откатить параметры предложения к указанной версии. Это считается новой правкой, поэтому версия инкрементируется.
      operationId: rollbackBid
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/bidId"
        - name: version
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Номер версии, к которой нужно откатить предложение.
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Предложение успешно откатано и версия инкрементирована.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bid"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Предложение или версия не найдены.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
...
components:
  schemas:
    username:
      type: string
      description: Уникальный slug пользователя.
      example: test_user
...
    tenderId:
      type: string
      description: Уникальный идентификатор тендера, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
...
    organizationId:
      type: string
      description: Уникальный идентификатор организации, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
...
    bidStatus:
      type: string
      description: Статус предложения
      enum:
        - Created
        - Published
        - Canceled
        - Approved
        - Rejected
    bidDecision:
      type: string
      description: Решение по предложению
      enum:
        - Approved
        - Rejected
    bidId:
      type: string
      description: Уникальный идентификатор предложения, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    bidName:
      type: string
      description: Полное название предложения
      maxLength: 100
    bidDescription:
      type: string
      description: Описание предложения
      maxLength: 500
...
    bidAuthorType:
      type: string
      description: Тип автора
      enum:
        - Organization
        - User
    bidAuthorId:
      type: string
      description: Уникальный идентификатор автора предложения, присвоенный сервером.
      example: 550e8400-e29b-41d4-a716-446655440000
      maxLength: 100
    bidVersion:
      type: integer
      description: Номер версии посел правок
      format: int32
      minimum: 1
      default: 1
...
    bid:
      type: object
      description: Информация о предложении
      properties:
        id:
          $ref: "#/components/schemas/bidId"
        name:
          $ref: "#/components/schemas/bidName"
        description:
          $ref: "#/components/schemas/bidDescription"
        status:
          $ref: "#/components/schemas/bidStatus"
        tenderId:
          $ref: "#/components/schemas/tenderId"
        authorType:
          $ref: "#/components/schemas/bidAuthorType"
        authorId:
          $ref: "#/components/schemas/bidAuthorId"
        version:
          $ref: "#/components/schemas/bidVersion"
        createdAt:
          type: string
          description: |
            Серверная дата и время в момент, когда пользователь отправил предложение на создание.
            Передается в формате RFC3339.
          example: 2006-01-02T15:04:05Z07:00
        
      required:
        - id
        - name
        - description
        - status
        - tenderId
        - createdAt
        - authorType
        - authorId
        - version
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        name: Доставка товаров Алексей
        status: Created
        authorType: User
        authorId: 61a485f0-e29b-41d4-a716-446655440000
        verstion: 1
        createdAt: 2006-01-02T15:04:05Z07:00
    errorResponse:
      type: object
      description: Используется для возвращения ошибки пользователю
      properties:
        reason:
          type: string
          description: Описание ошибки в свободной форме
          minLength: 5
      required:
        - reason
      example:
        reason: <объяснение, почему запрос пользователя не может быть обработан>
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer